+++ crates/client/rpc/src/versions/user/v0_7_1/methods/read/get_block_with_receipts.rs
@@ -248,6 +248,8 @@ mod tests {
                 },
                 StateDiff::default(),
                 vec![],
+                None,
+                None,
             )
             .unwrap();