+++ crates/client/rpc/src/versions/user/v0_7_1/methods/trace/trace_block_transactions.rs
@@ -8,7 +8,7 @@ use std::sync::Arc;
 
 use super::trace_transaction::FALLBACK_TO_SEQUENCER_WHEN_VERSION_BELOW;
 use crate::errors::{StarknetRpcApiError, StarknetRpcResult};
-use crate::utils::transaction::to_blockifier_transactions;
+use crate::utils::transaction::to_blockifier_transaction;
 use crate::utils::ResultExt;
 use crate::Starknet;
 
@@ -30,7 +30,7 @@ pub async fn trace_block_transactions(
         .into_iter()
         .zip(block.info.tx_hashes())
         .map(|(tx, hash)| {
-            to_blockifier_transactions(starknet.clone_backend(), block_id.clone(), tx, &TransactionHash(*hash))
+            to_blockifier_transaction(starknet.clone_backend(), block_id.clone(), tx, &TransactionHash(*hash))
         })
         .collect::<Result<_, _>>()?;