+++ crates/tests/src/lib.rs
@@ -6,10 +6,14 @@ mod storage_proof;
 
 use anyhow::bail;
 use rstest::rstest;
+use starknet::accounts::{Account, ExecutionEncoding, SingleOwnerAccount};
+use starknet::signers::{LocalWallet, SigningKey};
+use starknet_core::types::{BlockId, BlockTag, Call, Felt};
+use starknet_core::utils::starknet_keccak;
 use starknet_providers::Provider;
 use starknet_providers::{jsonrpc::HttpTransport, JsonRpcClient, Url};
-use std::ops::Range;
-use std::sync::Mutex;
+use std::ops::{Deref, Range};
+use std::sync::{Arc, Mutex};
 use std::{
     collections::HashMap,
     env,
@@ -44,9 +48,9 @@ async fn wait_for_cond<F: Future<Output = Result<(), anyhow::Error>>>(
 pub struct MadaraCmd {
     process: Option<Child>,
     ready: bool,
-    json_rpc: Option<JsonRpcClient<HttpTransport>>,
+    json_rpc: JsonRpcClient<HttpTransport>,
     rpc_url: Url,
-    tempdir: TempDir,
+    tempdir: Arc<TempDir>,
     _port: MadaraPortNum,
 }
 
@@ -55,8 +59,8 @@ impl MadaraCmd {
         self.process.take().unwrap().wait_with_output().unwrap()
     }
 
-    pub fn json_rpc(&mut self) -> &JsonRpcClient<HttpTransport> {
-        self.json_rpc.get_or_insert_with(|| JsonRpcClient::new(HttpTransport::new(self.rpc_url.clone())))
+    pub fn json_rpc(&self) -> &JsonRpcClient<HttpTransport> {
+        &self.json_rpc
     }
 
     pub fn db_dir(&self) -> &Path {
@@ -71,8 +75,8 @@ impl MadaraCmd {
                 res.error_for_status()?;
                 anyhow::Ok(())
             },
-            Duration::from_secs(2),
-            10,
+            Duration::from_millis(500),
+            20,
         )
         .await;
         self.ready = true;
@@ -146,6 +150,7 @@ lazy_static::lazy_static! {
     static ref AVAILABLE_PORTS: Mutex<AvailablePorts<Range<u16>>> = Mutex::new(AvailablePorts { to_reuse: vec![], next: PORT_RANGE });
 }
 
+#[derive(Clone)]
 pub struct MadaraPortNum(pub u16);
 impl Drop for MadaraPortNum {
     fn drop(&mut self) {
@@ -163,10 +168,11 @@ pub fn get_port() -> MadaraPortNum {
     MadaraPortNum(port)
 }
 
+#[derive(Clone)]
 pub struct MadaraCmdBuilder {
     args: Vec<String>,
     env: HashMap<String, String>,
-    tempdir: TempDir,
+    tempdir: Arc<TempDir>,
     port: MadaraPortNum,
 }
 
@@ -181,7 +187,7 @@ impl MadaraCmdBuilder {
         Self {
             args: Default::default(),
             env: Default::default(),
-            tempdir: TempDir::with_prefix("madara-test").unwrap(),
+            tempdir: Arc::new(TempDir::with_prefix("madara-test").unwrap()),
             port: get_port(),
         }
     }
@@ -213,7 +219,7 @@ impl MadaraCmdBuilder {
                     .into_iter()
                     .chain([
                         "--base-path".into(),
-                        format!("{}", self.tempdir.as_ref().display()),
+                        format!("{}", self.tempdir.deref().as_ref().display()),
                         "--rpc-port".into(),
                         format!("{}", self.port.0),
                     ])
@@ -223,11 +229,12 @@ impl MadaraCmdBuilder {
             .spawn()
             .unwrap();
 
+        let rpc_url = Url::parse(&format!("http://127.0.0.1:{}/", self.port.0)).unwrap();
         MadaraCmd {
             process: Some(process),
             ready: false,
-            json_rpc: None,
-            rpc_url: Url::parse(&format!("http://127.0.0.1:{}/", self.port.0)).unwrap(),
+            json_rpc: JsonRpcClient::new(HttpTransport::new(rpc_url.clone())),
+            rpc_url,
             tempdir: self.tempdir,
             _port: self.port,
         }
@@ -276,4 +283,199 @@ async fn madara_can_sync_a_few_blocks() {
             block_number: 19
         }
     );
+}
+
+const ERC20_STRK_CONTRACT_ADDRESS: Felt =
+    Felt::from_hex_unchecked("0x04718f5a0fc34cc1af16a1cdee98ffb20c31f5cd61d6ab07201858f4287c938d");
+#[allow(unused)]
+const ERC20_ETH_CONTRACT_ADDRESS: Felt =
+    Felt::from_hex_unchecked("0x049d36570d4e46f48e99674bd3fcc84644ddd6b96f7c741b1562b82f9e004dc7");
+
+const ACCOUNT_SECRET: Felt =
+    Felt::from_hex_unchecked("0x077e56c6dc32d40a67f6f7e6625c8dc5e570abe49c0a24e9202e4ae906abcc07");
+const ACCOUNT_ADDRESS: Felt =
+    Felt::from_hex_unchecked("0x055be462e718c4166d656d11f89e341115b8bc82389c3762a10eade04fcb225d");
+
+#[rstest]
+#[tokio::test]
+async fn madara_devnet_add_transaction() {
+    let _ = tracing_subscriber::fmt().with_test_writer().try_init();
+
+    let args = &[
+        "--devnet",
+        "--no-l1-sync",
+        "--gas-price",
+        "0",
+        // only produce blocks no pending txs
+        "--chain-config-override",
+        "block_time=1s,pending_block_update_time=1s",
+    ];
+
+    let cmd_builder = MadaraCmdBuilder::new().args(*args);
+    let mut node = cmd_builder.run();
+    node.wait_for_ready().await;
+
+    tokio::time::sleep(Duration::from_secs(3)).await;
+
+    let chain_id = node.json_rpc().chain_id().await.unwrap();
+
+    let signer = LocalWallet::from_signing_key(SigningKey::from_secret_scalar(ACCOUNT_SECRET));
+    let mut account =
+        SingleOwnerAccount::new(node.json_rpc(), signer, ACCOUNT_ADDRESS, chain_id, ExecutionEncoding::New);
+    account.set_block_id(BlockId::Tag(BlockTag::Latest));
+
+    let res = account
+        .execute_v3(vec![Call {
+            to: ERC20_STRK_CONTRACT_ADDRESS,
+            selector: starknet_keccak(b"transfer"),
+            calldata: vec![ACCOUNT_ADDRESS, 15.into(), Felt::ZERO],
+        }])
+        .send()
+        .await
+        .unwrap();
+
+    wait_for_cond(
+        || async {
+            let _receipt = node.json_rpc().get_transaction_receipt(res.transaction_hash).await?;
+            Ok(())
+        },
+        Duration::from_millis(500),
+        60,
+    )
+    .await;
+}
+
+#[rstest]
+#[tokio::test]
+async fn madara_devnet_mempool_saving() {
+    let _ = tracing_subscriber::fmt().with_test_writer().try_init();
+
+    let cmd_builder = MadaraCmdBuilder::new().args([
+        "--devnet",
+        "--no-l1-sync",
+        "--gas-price",
+        "0",
+        // never produce blocks & pending txs
+        "--chain-config-path",
+        "test_devnet.yaml",
+        "--chain-config-override",
+        "block_time=5min,pending_block_update_time=5min",
+    ]);
+    let mut node = cmd_builder.clone().run();
+    node.wait_for_ready().await;
+
+    let chain_id = node.json_rpc().chain_id().await.unwrap();
+
+    let signer = LocalWallet::from_signing_key(SigningKey::from_secret_scalar(ACCOUNT_SECRET));
+    let mut account =
+        SingleOwnerAccount::new(node.json_rpc(), signer, ACCOUNT_ADDRESS, chain_id, ExecutionEncoding::New);
+    account.set_block_id(BlockId::Tag(BlockTag::Pending));
+
+    let res = account
+        .execute_v3(vec![Call {
+            to: ERC20_STRK_CONTRACT_ADDRESS,
+            selector: starknet_keccak(b"transfer"),
+            calldata: vec![ACCOUNT_ADDRESS, 15.into(), Felt::ZERO],
+        }])
+        .send()
+        .await
+        .unwrap();
+
+    drop(node);
+
+    // tx should be in saved mempool
+
+    let cmd_builder = cmd_builder.args([
+        "--devnet",
+        "--no-l1-sync",
+        "--gas-price",
+        "0",
+        // never produce blocks but produce pending txs
+        "--chain-config-path",
+        "test_devnet.yaml",
+        "--chain-config-override",
+        "block_time=5min,pending_block_update_time=500ms",
+    ]);
+    let mut node = cmd_builder.clone().run();
+    node.wait_for_ready().await;
+
+    // tx should be in mempool
+
+    wait_for_cond(
+        || async {
+            let _receipt = node.json_rpc().get_transaction_receipt(res.transaction_hash).await?;
+            Ok(())
+        },
+        Duration::from_millis(500),
+        60,
+    )
+    .await;
+}
+
+#[rstest]
+#[tokio::test]
+async fn madara_devnet_continue_pending() {
+    let _ = tracing_subscriber::fmt().with_test_writer().try_init();
+
+    let cmd_builder = MadaraCmdBuilder::new().args([
+        "--devnet",
+        "--no-l1-sync",
+        "--gas-price",
+        "0",
+        // never produce blocks but produce pending txs
+        "--chain-config-path",
+        "test_devnet.yaml",
+        "--chain-config-override",
+        "block_time=5min,pending_block_update_time=500ms",
+    ]);
+    let mut node = cmd_builder.clone().run();
+    node.wait_for_ready().await;
+
+    let chain_id = node.json_rpc().chain_id().await.unwrap();
+
+    let signer = LocalWallet::from_signing_key(SigningKey::from_secret_scalar(ACCOUNT_SECRET));
+    let mut account =
+        SingleOwnerAccount::new(node.json_rpc(), signer, ACCOUNT_ADDRESS, chain_id, ExecutionEncoding::New);
+    account.set_block_id(BlockId::Tag(BlockTag::Pending));
+
+    let res = account
+        .execute_v3(vec![Call {
+            to: ERC20_STRK_CONTRACT_ADDRESS,
+            selector: starknet_keccak(b"transfer"),
+            calldata: vec![ACCOUNT_ADDRESS, 15.into(), Felt::ZERO],
+        }])
+        .send()
+        .await
+        .unwrap();
+
+    wait_for_cond(
+        || async {
+            let _receipt = node.json_rpc().get_transaction_receipt(res.transaction_hash).await?;
+            Ok(())
+        },
+        Duration::from_millis(500),
+        60,
+    )
+    .await;
+
+    drop(node);
+
+    // tx should appear in saved pending block
+
+    let cmd_builder = cmd_builder.args([
+        "--devnet",
+        "--no-l1-sync",
+        "--gas-price",
+        "0",
+        // never produce blocks never produce pending txs
+        "--chain-config-path",
+        "test_devnet.yaml",
+        "--chain-config-override",
+        "block_time=5min,pending_block_update_time=5min",
+    ]);
+    let mut node = cmd_builder.clone().run();
+    node.wait_for_ready().await;
+
+    // should find receipt
+    let _receipt = node.json_rpc().get_transaction_receipt(res.transaction_hash).await.unwrap();
 }